packages/dnsdump/pyproject.toml
[project]
name = "dnsdump"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Julien OPPLIGER", email = "account@oppliger.cc" }
]
requires-python = ">=3.13"
dependencies = [
    "apex",
    "common",
    "dnspython>=2.8.0",
    "open-asset-model>=1.1.1",
    "open-asset-store>=0.0.2",
    "pygments>=2.19.2",
    "pyrate-limiter>=4.0.2",
    "python-dotenv>=1.2.1",
    "termcolor>=3.3.0",
]

[build-system]
requires = ["uv_build>=0.9.21,<0.10.0"]
build-backend = "uv_build"

[tool.uv.sources]
common = { workspace = true }
apex = { workspace = true }

[project.scripts]
dnsdump = "dnsdump.__main__:main"
